feat(market): move market into application module
