+++ src/game.cpp
@@ -14,7 +14,7 @@
 #include "housetile.h"
 #include "http/http.h"
 #include "iologindata.h"
-#include "iomarket.h"
+#include "modules/market/module.h"
 #include "items.h"
 #include "movement.h"
 #include "outfit.h"
@@ -4935,8 +4935,8 @@ void Game::playerBrowseMarket(uint32_t playerId, uint16_t spriteId)
 		return;
 	}
 
-	const MarketOfferList& buyOffers = tfs::iomarket::getActiveOffers(MARKETACTION_BUY, it.id);
-	const MarketOfferList& sellOffers = tfs::iomarket::getActiveOffers(MARKETACTION_SELL, it.id);
+	const MarketOfferList& buyOffers = tfs::modules::market::getActiveOffers(MARKETACTION_BUY, it.id);
+	const MarketOfferList& sellOffers = tfs::modules::market::getActiveOffers(MARKETACTION_SELL, it.id);
 	player->sendMarketBrowseItem(it.id, buyOffers, sellOffers);
 	tfs::events::player::onLookInMarket(player, &it);
 }
@@ -4945,8 +4945,10 @@ void Game::playerBrowseMarketOwnOffers(uint32_t playerId)
 {
 	if (const auto& player = getPlayerByID(playerId)) {
 		if (player->isInMarket()) {
-			const MarketOfferList& buyOffers = tfs::iomarket::getOwnOffers(MARKETACTION_BUY, player->getGUID());
-			const MarketOfferList& sellOffers = tfs::iomarket::getOwnOffers(MARKETACTION_SELL, player->getGUID());
+			const MarketOfferList& buyOffers =
+			    tfs::modules::market::getOwnOffers(MARKETACTION_BUY, player->getGUID());
+			const MarketOfferList& sellOffers =
+			    tfs::modules::market::getOwnOffers(MARKETACTION_SELL, player->getGUID());
 			player->sendMarketBrowseOwnOffers(buyOffers, sellOffers);
 		}
 	}
@@ -4956,9 +4958,10 @@ void Game::playerBrowseMarketOwnHistory(uint32_t playerId)
 {
 	if (const auto& player = getPlayerByID(playerId)) {
 		if (player->isInMarket()) {
-			const HistoryMarketOfferList& buyOffers = tfs::iomarket::getOwnHistory(MARKETACTION_BUY, player->getGUID());
+			const HistoryMarketOfferList& buyOffers =
+			    tfs::modules::market::getOwnHistory(MARKETACTION_BUY, player->getGUID());
 			const HistoryMarketOfferList& sellOffers =
-			    tfs::iomarket::getOwnHistory(MARKETACTION_SELL, player->getGUID());
+			    tfs::modules::market::getOwnHistory(MARKETACTION_SELL, player->getGUID());
 			player->sendMarketBrowseOwnHistory(buyOffers, sellOffers);
 		}
 	}
@@ -5008,7 +5011,7 @@ void Game::playerCreateMarketOffer(uint32_t playerId, uint8_t type, uint16_t spr
 	}
 
 	const uint32_t maxOfferCount = getNumber(ConfigManager::MAX_MARKET_OFFERS_AT_A_TIME_PER_PLAYER);
-	if (maxOfferCount != 0 && tfs::iomarket::getPlayerOfferCount(player->getGUID()) >= maxOfferCount) {
+	if (maxOfferCount != 0 && tfs::modules::market::getPlayerOfferCount(player->getGUID()) >= maxOfferCount) {
 		return;
 	}
 
@@ -5064,11 +5067,12 @@ void Game::playerCreateMarketOffer(uint32_t playerId, uint8_t type, uint16_t spr
 		player->setBankBalance(player->getBankBalance() - debitBank);
 	}
 
-	tfs::iomarket::createOffer(player->getGUID(), static_cast<MarketAction_t>(type), it.id, amount, price, anonymous);
+	tfs::modules::market::createOffer(
+	    player->getGUID(), static_cast<MarketAction_t>(type), it.id, amount, price, anonymous);
 
 	player->sendMarketEnter();
-	const MarketOfferList& buyOffers = tfs::iomarket::getActiveOffers(MARKETACTION_BUY, it.id);
-	const MarketOfferList& sellOffers = tfs::iomarket::getActiveOffers(MARKETACTION_SELL, it.id);
+	const MarketOfferList& buyOffers = tfs::modules::market::getActiveOffers(MARKETACTION_BUY, it.id);
+	const MarketOfferList& sellOffers = tfs::modules::market::getActiveOffers(MARKETACTION_SELL, it.id);
 	player->sendMarketBrowseItem(it.id, buyOffers, sellOffers);
 }
 
@@ -5083,7 +5087,7 @@ void Game::playerCancelMarketOffer(uint32_t playerId, uint32_t timestamp, uint16
 		return;
 	}
 
-	MarketOfferEx offer = tfs::iomarket::getOfferByCounter(timestamp, counter);
+	MarketOfferEx offer = tfs::modules::market::getOfferByCounter(timestamp, counter);
 	if (offer.id == 0 || offer.playerId != player->getGUID()) {
 		return;
 	}
@@ -5125,7 +5129,7 @@ void Game::playerCancelMarketOffer(uint32_t playerId, uint32_t timestamp, uint16
 		}
 	}
 
-	tfs::iomarket::moveOfferToHistory(offer.id, OFFERSTATE_CANCELLED);
+	tfs::modules::market::moveOfferToHistory(offer.id, OFFERSTATE_CANCELLED);
 	offer.amount = 0;
 	offer.timestamp += getNumber(ConfigManager::MARKET_OFFER_DURATION);
 	player->sendMarketCancelOffer(offer);
@@ -5147,7 +5151,7 @@ void Game::playerAcceptMarketOffer(uint32_t playerId, uint32_t timestamp, uint16
 		return;
 	}
 
-	MarketOfferEx offer = tfs::iomarket::getOfferByCounter(timestamp, counter);
+	MarketOfferEx offer = tfs::modules::market::getOfferByCounter(timestamp, counter);
 	if (offer.id == 0) {
 		return;
 	}
@@ -5286,19 +5290,20 @@ void Game::playerAcceptMarketOffer(uint32_t playerId, uint32_t timestamp, uint16
 
 	const int32_t marketOfferDuration = getNumber(ConfigManager::MARKET_OFFER_DURATION);
 
-	tfs::iomarket::appendHistory(player->getGUID(),
-	                             (offer.type == MARKETACTION_BUY ? MARKETACTION_SELL : MARKETACTION_BUY), offer.itemId,
-	                             amount, offer.price, offer.timestamp + marketOfferDuration, OFFERSTATE_ACCEPTEDEX);
+	tfs::modules::market::appendHistory(
+	    player->getGUID(), (offer.type == MARKETACTION_BUY ? MARKETACTION_SELL : MARKETACTION_BUY), offer.itemId,
+	    amount, offer.price, offer.timestamp + marketOfferDuration, OFFERSTATE_ACCEPTEDEX);
 
-	tfs::iomarket::appendHistory(offer.playerId, offer.type, offer.itemId, amount, offer.price,
-	                             offer.timestamp + marketOfferDuration, OFFERSTATE_ACCEPTED);
+	tfs::modules::market::appendHistory(
+	    offer.playerId, offer.type, offer.itemId, amount, offer.price, offer.timestamp + marketOfferDuration,
+	    OFFERSTATE_ACCEPTED);
 
 	offer.amount -= amount;
 
 	if (offer.amount == 0) {
-		tfs::iomarket::deleteOffer(offer.id);
+		tfs::modules::market::deleteOffer(offer.id);
 	} else {
-		tfs::iomarket::acceptOffer(offer.id, amount);
+		tfs::modules::market::acceptOffer(offer.id, amount);
 	}
 
 	player->sendMarketEnter();
