+++ src/modules/market/module.h
@@ -1,20 +1,25 @@
 // Copyright 2023 The Forgotten Server Authors. All rights reserved.
 // Use of this source code is governed by the GPL-2.0 License that can be found in the LICENSE file.
 
-#ifndef FS_IOMARKET_H
-#define FS_IOMARKET_H
+#ifndef FS_MODULES_MARKET_MODULE_H
+#define FS_MODULES_MARKET_MODULE_H
 
-#include "database.h"
-#include "enums.h"
+#include "../../application/module.h"
+#include "resources.h"
 
-namespace tfs::iomarket {
+#include <ctime>
+
+namespace tfs::modules::market {
+
+class MarketModule final : public application::Module
+{
+public:
+	void build(application::App& app) override;
+};
 
 MarketOfferList getActiveOffers(MarketAction_t action, uint16_t itemId);
 MarketOfferList getOwnOffers(MarketAction_t action, uint32_t playerId);
 HistoryMarketOfferList getOwnHistory(MarketAction_t action, uint32_t playerId);
-
-void processExpiredOffers(std::shared_ptr<DBResult> result, bool);
-void checkExpiredOffers();
 
 uint32_t getPlayerOfferCount(uint32_t playerId);
 MarketOfferEx getOfferByCounter(uint32_t timestamp, uint16_t counter);
@@ -28,11 +33,9 @@ void appendHistory(uint32_t playerId, MarketAction_t type, uint16_t itemId, uint
                    time_t timestamp, MarketOfferState_t state);
 bool moveOfferToHistory(uint32_t offerId, MarketOfferState_t state);
 
-void updateStatistics();
+MarketStatistics* getPurchaseStatistics(MarketPurchaseStatistics& purchaseStatistics, uint16_t itemId);
+MarketStatistics* getSaleStatistics(MarketSaleStatistics& saleStatistics, uint16_t itemId);
 
-MarketStatistics* getPurchaseStatistics(uint16_t itemId);
-MarketStatistics* getSaleStatistics(uint16_t itemId);
+} // namespace tfs::modules::market
 
-} // namespace tfs::iomarket
-
-#endif // FS_IOMARKET_H
+#endif // FS_MODULES_MARKET_MODULE_H
