+++ vc18/atlas.vcxproj
@@ -157,7 +157,6 @@
     <ClCompile Include="..\src\iologindata.cpp" />
     <ClCompile Include="..\src\iomap.cpp" />
     <ClCompile Include="..\src\iomapserialize.cpp" />
-    <ClCompile Include="..\src\iomarket.cpp" />
     <ClCompile Include="..\src\item.cpp" />
     <ClCompile Include="..\src\items.cpp" />
     <ClCompile Include="..\src\lua\api.cpp" />
@@ -228,6 +227,7 @@
     <ClCompile Include="..\src\rsa.cpp" />
     <ClCompile Include="..\src\app_loop.cpp" />
     <ClCompile Include="..\src\modules\app_loop\module.cpp" />
+    <ClCompile Include="..\src\modules\market\module.cpp" />
     <ClCompile Include="..\src\modules\server\module.cpp" />
     <ClCompile Include="..\src\script.cpp" />
     <ClCompile Include="..\src\scriptmanager.cpp" />
@@ -305,7 +305,6 @@
     <ClInclude Include="..\src\iologindata.h" />
     <ClInclude Include="..\src\iomap.h" />
     <ClInclude Include="..\src\iomapserialize.h" />
-    <ClInclude Include="..\src\iomarket.h" />
     <ClInclude Include="..\src\item.h" />
     <ClInclude Include="..\src\itemloader.h" />
     <ClInclude Include="..\src\items.h" />
@@ -342,6 +341,8 @@
     <ClInclude Include="..\src\rsa.h" />
     <ClInclude Include="..\src\app_loop.h" />
     <ClInclude Include="..\src\modules\app_loop\module.h" />
+    <ClInclude Include="..\src\modules\market\module.h" />
+    <ClInclude Include="..\src\modules\market\resources.h" />
     <ClInclude Include="..\src\modules\server\module.h" />
     <ClInclude Include="..\src\script.h" />
     <ClInclude Include="..\src\scriptmanager.h" />
